requirements.txt
asgiref==3.5.0
astral==2.2
autopep8==1.6.0
beautiful-date==2.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
Django==4.0.4
django-multiselectfield==0.1.12
et-xmlfile==1.1.0
fonttools==4.32.0
gcsa==1.3.0
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.1
meteostat==1.6.1
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
Pebble==4.6.3
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.9